added get function in savedCountriesList
